+++ examples/top_lite_simulator.py
@@ -1,7 +1,7 @@
 """Lite simulation of the top linux command."""
-
 import datetime
 import random
+import sys
 import time
 from dataclasses import dataclass
 
@@ -10,10 +10,10 @@ from rich.console import Console
 from rich.live import Live
 from rich.table import Table
 
-try:
-    from typing_extensions import Literal
-except ImportError:
+if sys.version_info >= (3, 8):
     from typing import Literal
+else:
+    from typing_extensions import Literal
 
 
 @dataclass