+++ rich/_timer.py
@@ -6,10 +6,11 @@ Timer context manager, only used in debug.
 from time import time
 
 import contextlib
+from typing import Generator
 
 
 @contextlib.contextmanager
-def timer(subject: str = "time"):
+def timer(subject: str = "time") -> Generator[None, None, None]:
     """print the elapsed time. (only used in debugging)"""
     start = time()
     yield