+++ rich/color.py
@@ -305,7 +305,7 @@ class Color(NamedTuple):
         return self.type == ColorType.DEFAULT
 
     def get_truecolor(
-        self, theme: Optional["TerminalTheme"] = None, foreground=True
+        self, theme: Optional["TerminalTheme"] = None, foreground: bool = True
     ) -> ColorTriplet:
         """Get an equivalent color triplet for this color.
 
@@ -471,7 +471,7 @@ class Color(NamedTuple):
     def downgrade(self, system: ColorSystem) -> "Color":
         """Downgrade a color system to a system with fewer colors."""
 
-        if self.type == ColorType.DEFAULT or self.type == system:
+        if self.type in [ColorType.DEFAULT, system]:
             return self
         # Convert to 8-bit color from truecolor color
         if system == ColorSystem.EIGHT_BIT and self.system == ColorSystem.TRUECOLOR: