+++ rich/measure.py
@@ -1,5 +1,5 @@
 from operator import itemgetter
-from typing import Iterable, NamedTuple, Optional, TYPE_CHECKING
+from typing import Callable, Iterable, NamedTuple, Optional, TYPE_CHECKING
 
 from . import errors
 from .protocol import is_renderable
@@ -100,7 +100,9 @@ class Measurement(NamedTuple):
         if hasattr(renderable, "__rich__"):
             renderable = renderable.__rich__()  # type: ignore
         if is_renderable(renderable):
-            get_console_width = getattr(renderable, "__rich_measure__", None)
+            get_console_width: Callable[
+                ["Console", "ConsoleOptions"], "Measurement"
+            ] = getattr(renderable, "__rich_measure__", None)
             if get_console_width is not None:
                 render_width = (
                     get_console_width(console, options)