+++ rich/pager.py
@@ -1,4 +1,5 @@
 from abc import ABC, abstractmethod
+from typing import Any, Callable
 
 
 class Pager(ABC):
@@ -16,7 +17,9 @@ class Pager(ABC):
 class SystemPager(Pager):
     """Uses the pager installed on the system."""
 
-    _pager = lambda self, content: __import__("pydoc").pager(content)
+    _pager: Callable[[Any, str], Any] = lambda self, content: __import__("pydoc").pager(
+        content
+    )
 
     def show(self, content: str) -> None:
         """Use the same pager used by pydoc."""